+++ docs/changes.rst
@@ -17,7 +17,7 @@ Unreleased.
  - Added :meth:`Image.annotate() <wand.image.BaseImage.annotate>` method. [:issue:`#418`]
  - Added :meth:`Image.percent_escape() <wand.image.BaseImage.percent_escape>` helper method. [:issue:`421`]
  - Added :meth:`Image.ping() <wand.image.Image.ping>` class method. [:issue:`#425`]
- 
+ - Added ``mean_color``, ``keep``, & ``remove`` parameters in :meth:`Image.connected_components() <wand.image.BaseImage.connected_components>` method.
 
 .. _changelog-0.5.5: